% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuriR.R
\docType{data}
\name{cols}
\alias{cols}
\title{Yuri}
\format{
An object of class \code{character} of length 12.
}
\usage{
cols
}
\arguments{
\item{cols}{}

\item{fills}{}

\item{t_theme}{}

\item{Yuri_theme}{}

\item{outer.IQR}{}

\item{Up}{}
}
\description{
cols, fills, t_theme, Yuri_theme, outer.IQR, Up
}
\examples{

}
\keyword{datasets}
