% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_theme.r
\docType{data}
\name{lin_theme}
\alias{lin_theme}
\alias{col1}
\title{lin_theme}
\format{
An object of class \code{character} of length 7.
}
\usage{
col1
}
\description{
lin_theme, cols, fills
}
\keyword{datasets}
