% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.r
\name{dca}
\alias{dca}
\title{dca}
\usage{
dca(
  data,
  outcome,
  predictors,
  xstart = 0.01,
  xstop = 0.99,
  xby = 0.01,
  ymin = -0.05,
  probability = NULL,
  harm = NULL,
  graph = TRUE,
  intervention = FALSE,
  interventionper = 100,
  smooth = FALSE,
  loess.span = 0.1
)
}
\description{
dca
}
