% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linr.R
\docType{package}
\name{linr}
\alias{linr}
\title{Function to Fit Linear Regression}
\usage{
linr(formula, data, method = "cholesky")
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted with specific pattern (i.e. y ~ x1 + x2 where y is the responding variable; x1 and x2 are the covariates for the model)}

\item{data}{an optional data frame, list or environment containing the variables in the model. If not found in data, the variables are taken from formula.}

\item{method}{an optional character string specifying the fitting method. It must be one of the strings "qr", "cholesky", or "svd".}
}
\value{
A list containing the following elements
        \itemize{
          \item call - return the fitted linear model fomula and the corresponding data.
          \item coefficients - a table with the estimated values for each covariates and the intercept as well as their corresponding standard error, t-statistics and (two-sided) p-value.
          \item residuals - summary of the usual residuals with min, 1st quantile, median, 3rd quantile, max being computed.
          \item fitted.values - the fitted mean values.
          \item Residual standard error - the square root of the estimated variance of the random error and a corresponding degrees of freedom will also be computed.
          \item R.squared - R^2, the 'fraction of variance explained by the model', SSR (variation in fitted values about the overall mean) / SSY (total variation in Y about its overall mean).
          \item Adj.R.squared - adjusted version of R^2, penalized based on number of covariates.
          \item SE.beta - the corresponding standard error for the estimated coefficient.
          \item t value - the t-statistic for corresponding variable.
          \item Pr(>|t|) - the corresponding (two-sided) p-value for the t-statistic
          \item F.statistic - the test statistic for F-tests. Variation Between Sample Means / Variation Within the Samples.
        }
}
\description{
"linr" is used to fit a simple linear model. In this function, linear regression can be done by three matrix decomposition methods, which are the QR decomposition, Cholesky decomposition and the singular value decomposition (SVD). The defalt fitting method used is the QR method which is used in "lm" function in "stats" package. All three methods can fit linear model with high efficiency.
}
\examples{
Linear_model.lm <- lm(dist~speed,data=cars)
print(Linear_model.lm)
Linear_model.linr <- linr(dist~speed,data=cars)
print(Linear_model.linr$Call)
print(Linear_model.linr$Coefficients)

}
\seealso{
\code{\link[utils]{head}}
}
