% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdca.r
\name{stdca}
\alias{stdca}
\title{stdca}
\usage{
stdca(
  data,
  outcome,
  ttoutcome,
  timepoint,
  predictors,
  xstart = 0.001,
  xstop = 0.999,
  xby = 0.001,
  ymin = -0.05,
  probability = NULL,
  harm = NULL,
  intervention = F,
  interventionper = 100,
  smooth = T,
  loess.span = 0.1,
  cmprsk = F
)
}
\description{
stdca
}
